Fix stale links to Tails files, version 0.20.1 in create-image.sh
File:   create-image.sh

Update create-image.sh to point to tails 0.20.1 release.

The Tails .iso and consequently, it's GPG signature file, have been
updated to version 0.20.1. This patch simply updates the URL's to point
to the correct locations at the Tails repositories.

Without this patch create-image.sh is broken as of September 28, 2013.
With these changes things work as expected.
(Tails has moved the links to version 0.20.)